feat: created Constants file
I realized that when I changed my avatar image I forgot to update the path in the two spots where it's referenced and only did it in one. With this commit I've added a `Constants.ts` file where I added a `AvatarPath` variable that points to the path where the image can be found. Now, if I ever change the image later, I just need to update this one constant.